Move to doubles for the returns.

+++ exec/me_astro.h
@@ -4,13 +4,13 @@
 #define ME_ASTRO_H
 
 /* me_findlun.c */
-extern void me_findlun(long int mjd, long int mpm, float *ra, float *dec, float *dist);
+extern void me_findlun(long int mjd, long int mpm, double *ra, double *dec, double *dist);
 
 /* me_findjov.c */
-extern void me_findjov(long int mjd, long int mpm, float *ra, float *dec, float *dist);
+extern void me_findjov(long int mjd, long int mpm, double *ra, double *dec, double *dist);
 
 /* me_findsol.c */
-extern void me_findsol(long int mjd, long int mpm, float *ra, float *dec, float *dist);
+extern void me_findsol(long int mjd, long int mpm, double *ra, double *dec, double *dist);
 
 /* me_getaltaz.c */
 extern void me_getaltaz(double ra, double dec, double dist, long int mjd, long int mpm, double lat, double lon, double elev, double *LAST, double *alt, double *az);
